Moved KDTreeLinker(Algo|Tools).h from plugins to interface directory.
So that they can be used from both plugins and src directories.
