+++ RecoPixelVertexing/PixelTriplets/interface/KDTreeLinkerTools.h
