+++ RecoPixelVertexing/PixelTriplets/plugins/PixelTripletHLTGenerator.cc
@@ -18,8 +18,8 @@
 #include "DataFormats/GeometryVector/interface/Pi.h"
 //#include "RecoParticleFlow/PFProducer/interface/KDTreeLinkerAlgo.h"
 //#include "RecoParticleFlow/PFProducer/interface/KDTreeLinkerTools.h"
-#include "RecoPixelVertexing/PixelTriplets/plugins/KDTreeLinkerAlgo.h" //amend to point at your copy...
-#include "RecoPixelVertexing/PixelTriplets/plugins/KDTreeLinkerTools.h"
+#include "RecoPixelVertexing/PixelTriplets/interface/KDTreeLinkerAlgo.h" //amend to point at your copy...
+#include "RecoPixelVertexing/PixelTriplets/interface/KDTreeLinkerTools.h"
 
 using pixelrecoutilities::LongitudinalBendingCorrection;
 typedef PixelRecoRange<float> Range;
